Refactor: move entitie files
