+++ apps/api-server/src/auth/auth.controller.spec.ts
@@ -5,7 +5,7 @@ import { AuthService } from './auth.service';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
 import { MockUsersRepository } from '../../test/mock/user.repository.mock';
-import { User } from './entities/users.entity';
+import { User } from '../entities/users.entity';
 import { getRepositoryToken } from '@nestjs/typeorm';
 
 describe('AuthController', () => {
