+++ apps/api-server/src/database/database.config.ts
@@ -1,8 +1,8 @@
 import * as dotenv from 'dotenv';
 
-import { DrinksCategory } from '../drinks-category/drinks-category.entity';
+import { DrinksCategory } from '../entities/drinks-category.entity';
 import { SnakeNamingStrategy } from 'typeorm-naming-strategies';
-import { User } from '@src/auth/entities/users.entity';
+import { User } from '@src/entities/users.entity';
 import { databaseConfig } from '../config/config.constant';
 
 const dbConfig = databaseConfig().databaseConfig;
