+++ apps/api-server/src/drinks-category/drinks-category.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { DrinksCategory } from './drinks-category.entity';
+import { DrinksCategory } from '../entities/drinks-category.entity';
 
 @Injectable()
 export class DrinksCategoryService {
