+++ apps/api-server/src/users/users.controller.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 
 import { ConfigService } from '@nestjs/config';
 import { MockUsersRepository } from '../../test/mock/user.repository.mock';
-import { User } from '../auth/entities/users.entity';
+import { User } from '../entities/users.entity';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { getRepositoryToken } from '@nestjs/typeorm';
