+++ apps/api-server/src/users/users.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, Req, UseGuards } from '@nestjs/common';
-import { User } from '../auth/entities/users.entity';
+import { User } from '../entities/users.entity';
 import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
 import { UsersService } from './users.service';
