+++ apps/api-server/src/users/users.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { User } from '../auth/entities/users.entity';
+import { User } from '../entities/users.entity';
 
 @Injectable()
 export class UsersService {
